cauazinho_final_corrigido/cauazinho_agent/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
twilio==9.0.2
groq==0.5.0
newsapi-python==0.2.7
yfinance==0.2.37
gunicorn==21.2.0
requests==2.31.0
textblob==0.17.1
pandas==2.2.1
numpy==1.26.4
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.10
pytz==2024.1
scipy==1.12.0
pydantic==2.6.3
pydantic-core==2.16.3
openai==1.13.3
typing-extensions==4.10.0
httpx==0.27.0
httpcore==1.0.4
h11==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
werkzeug==3.0.1
blinker==1.7.0
